Make board generate on first click
Before, board was initialized at the start of a game. This caused some
first clicks to only reveal a single number cell or a bomb. Changed
this so that the board is generated when the first cell is revealed, so
that cell is guaranteed to be a non-bomb, non-number (number 0) cell.