Make frame and board modifiers declarative
Current functions for modifying the global frame and board modify the
global variables directly. Modified these functions so that they take
the current frame and board as parameters and return the modified
values. Also refactored frame to global_frame and board to global_board.

Now, these modifiers can be used in sub-processes, such as algorithms
for solving a minesweeper game.